+++ src/features/environments/components/index.tsx
@@ -1,5 +1,5 @@
-export * from "./PackageManager";
-export * from "./PackageManagerSearch";
+export * from "./Environments";
+export * from "./EnvironmentsSearch";
 export * from "./EnvironmentDropdown";
 export * from "./Environment";
 export * from "./EnvironmentsList";
